Update video slug
I finally give up on having the same URL structure for both videos and
playlists, it is definitely not possible without a lot of logic that
WordPress won’t easily let me do.

So instead I’m changing my URL structure for the video plugin.

* /%root%/%playlist%/ for the playlist page.
* /%root%/%playlist%/%prefix%/%video%/ for the video page.

Where %root% is the video slug defined in settings page (usually video),
%prefix% is the preffix that goes between playlist and video
(usually episode) and %playlist% and %video% are the slugs for playlist
and video.

As an example, assuming the default %root% and %prefix% values.
* http://www.example.com/videos/webinars/ for the playlist Webinars.
* http://www.example.com/videos/webinars/episode/making-money/ for a
  video named Making money in a playlist named Webinar.

The settings page has been modified too to make it possible to change
the prefix value.