Introduce new rewriting rules
The following rewriting rules have been introduced.

* Rewrite rule for the video archive.
* Rewrite rule for the playlist page: Playlist taxonomy has been
  modified so that slugs can be hierarchical now. This means that if
  a playlist is child of another taxonomy, it will appear in the URL
  as well.
* Rewrite rule for the single video. Video slug now renders the
  entire playlist slug including parent categories.

As an example, say that the user has a playlist named `Playlist`, 
which is child of `SuperPlaylist`, and parent of `SubPlaylist`. The
URL for listing videos in SubPlaylist will be:

http://www.example.com/videos/superplaylist/playlist/subplaylist

Say that there is a video in SubPlaylist. The complete slug will be:

/videos/superplaylist/playlist/subplaylist/episode/my-video/

These URLs are longer, but they actually give the user a clue of
which video in which playlists he is inspecting. On a real page,
where playlists are grouped in root playlists, this makes more sense.